refactor: machine.js 에서 validate (에러 검증) 함수들 분리하기
